+++ src/components/study-group-detail/index.ts
@@ -3,9 +3,8 @@ export { default as StudyGroupInfo } from '@components/study-group-detail/StudyG
 export { default as StudyGroupLecture } from '@components/study-group-detail/StudyGroupLecture'
 export { default as StudyGroupMember } from '@components/study-group-detail/StudyGroupMember'
 export { default as StudyGroupSchedule } from '@components/study-group-detail/StudyGroupSchedule'
-export { default as ScheduleCalendar } from '@/components/study-group-detail/Calendar/ScheduleCalendar'
-export { default as CalendarHeader } from '@components/study-group-detail/Calendar/CalendarHeader'
-export { default as CalendarDays } from '@components/study-group-detail/Calendar/CalendarDays'
-export { default as CalendarBody } from '@components/study-group-detail/Calendar/CalendarBody'
-export { default as CalendarBodyDate } from '@components/study-group-detail/Calendar/CalendarBodyDate'
-export { default as CalendarBodySchedule } from '@components/study-group-detail/Calendar/CalendarBodySchedule'
+export { default as ScheduleCalendar } from '@components/study-group-detail/calendar/ScheduleCalendar'
+export { default as CalendarHeader } from '@components/study-group-detail/calendar/CalendarHeader'
+export { default as CalendarDays } from '@components/study-group-detail/calendar/CalendarDays'
+export { default as CalendarBody } from '@components/study-group-detail/calendar/CalendarBody'
+export { default as CalendarBodyDate } from '@components/study-group-detail/calendar/CalendarBodyDate'
